Rename TrigPoly.h -> trig_poly.h
