+++ drake/util/test/testTrigPoly.cpp
@@ -1,4 +1,4 @@
-#include "drake/util/TrigPoly.h"
+#include "drake/util/trig_poly.h"
 
 #include <sstream>
 #include <map>
