+++ drake/util/trig_poly.h
